3.LLMEngine/requirements.txt
annotated-types==0.7.0
anyio==4.6.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.10
langchain-openai==0.2.2
langsmith==0.1.134
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
openai==1.51.2
orjson
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.39.2
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
